Create a Sankey plot using ggplot and ggalluvial

This function creates a Sankey plot with the help of ggplot and ggalluvial.

sankey_plot(
  df,
  left_var,
  right_var,
  xlab_setting,
  ylab_setting,
  name_left,
  name_right,
  title,
  title_size = 20,
  title_font = "verdana"
)

Arguments

df

A data frame containing the data to be plotted.

left_var

A string specifying the column name to be used as the left variable.

right_var

A string specifying the column name to be used as the right variable.

xlab_setting

ggplot labels settings for x axes.

ylab_setting

ggplot labels settings for y axes.

name_left

A string specifying the name for the left side of the plot.

name_right

A string specifying the name for the right side of the plot.

title

A string specifying the title of the plot.

title_size

Numeric value specifying the size of the title.

title_font

A string specifying the font of the title.

Value

A Sankey plot.